Updated receiver and sender code
Significantly updated the touch designer file, this is still in "R&D"
mode don't expect a ready to use component in your own projects ....
yet!

As for the nodeMCU code.

This is much more usable, but a warning:
There's a bug currently somewhere, may be voltage related that causes
the nodeMCU to while attempting to log into wifi or while receiving it's
first few frames of data.

However with a good router I am able to transmit via udp unicast to 8-10
devices at 60 fps.