Signed-off-by: Mario Bianchi <mariodbianchi@hotmail.com>
Rewriten effects loop and blink definition to implement a state machine to be able to interrupt effect by receiving new POST request
This may be a more effective effects engine, as now the frame to show will depend on the frame number.
So the effect is called and depending on the frame number you can calculate what is the frame to show.

Fixes:
1) New POST request overwrites current running effect.
2) While playing effects, all udp received packets are consumed in order to avoid "fast forward" effect and issue that prevented to receive new REST request while receiveing UDP packets at the same time